requirements.txt
Django==1.11.5
MySQL-python==1.2.5
Pillow==4.3.0
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.07.27.1
chardet==3.0.4
distribute==0.7.3
django-countries==4.6.2
django-extensions==2.1.0
django-mail-queue==3.1.1
django-sslserver==0.20
django-tinymce==2.7.0
et-xmlfile==1.0.1
idna==2.6
jdcal==1.3
kombu==4.1.0
meld3==1.0.2
olefile==0.44
openpyxl==2.4.8
pygraphviz==1.3.1
pytz==2017.2
requests==2.18.4
six==1.11.0
supervisor==3.3.4
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
wsgiref==0.1.2
